refactor(topic-right-pane): expose live tree utility
