+++ src/renderer/components/chat/flow/index.ts
@@ -1,8 +1,6 @@
 export { default as TopicMessageFlowCanvas } from './TopicMessageFlowCanvas'
 export { buildTopicMessageFlowGraph } from './topicMessageFlowGraph'
 export { layoutTopicMessageFlowGraph, TOPIC_MESSAGE_FLOW_NODE_SIZE } from './topicMessageFlowLayout'
-export type { TopicMessageFlowLiveNode, TopicMessageFlowLiveState } from './topicMessageFlowLiveTree'
-export { buildTopicMessageFlowLiveState, mergeTopicMessageFlowLiveTree } from './topicMessageFlowLiveTree'
 export { default as TopicMessageFlowNode } from './TopicMessageFlowNode'
 export type {
   TopicMessageFlowEdgeData,
