+++ src/renderer/pages/home/useChatRuntimeState.ts
@@ -1,10 +1,5 @@
 import { useInvalidateCache } from '@data/hooks/useDataApi'
 import { loggerService } from '@logger'
-// eslint-disable-next-line barrel/closed -- Bypass the flow barrel so chat startup does not touch TopicMessageFlowCanvas.
-import {
-  buildTopicMessageFlowLiveState,
-  type TopicMessageFlowLiveState
-} from '@renderer/components/chat/flow/topicMessageFlowLiveTree'
 import {
   type TranslationOverlayEntry,
   type TranslationOverlaySetter
@@ -21,6 +16,10 @@ import { useToolApprovalBridge } from '@renderer/hooks/useToolApprovalBridge'
 import { useTopicOverlayHandoffOnTerminal } from '@renderer/hooks/useTopicStreamStatus'
 import type { Topic } from '@renderer/types/topic'
 import { mergeMessagesById } from '@renderer/utils/message/mergeMessagesById'
+import {
+  buildTopicMessageFlowLiveState,
+  type TopicMessageFlowLiveState
+} from '@renderer/utils/topicMessageFlowLiveTree'
 import type { ActiveExecution } from '@shared/ai/transport'
 import type { CherryMessagePart, CherryUIMessage } from '@shared/data/types/message'
 import type { UniqueModelId } from '@shared/data/types/model'
